+++ nix/openbsd.nix
@@ -244,6 +244,7 @@ let
     chmod u+w "$out/miniaudio.h"
     patch -d "$out" -p1 < ${./miniaudio-openbsd-audio4.patch}
     patch -d "$out" -p1 < ${./miniaudio-openbsd35-headers.patch}
+    patch -d "$out" -p1 < ${./miniaudio-sndio-transfers.patch}
   '';
   av = pkgs.stdenvNoCC.mkDerivation {
     pname = "ffmpeg-headless-openbsd-${osVersion}";